Successfully sent a file between nodes
I have two separate directories for the 2 clients. One with the .torrent file and the completed file, and one with just the torrent file. The client with the completed file calls add_torrent() and begins seeding. The client without the file begins checks its dht for any nodes seeding and finds the other node, downloads all the pieces, and puts them together.