Gossip working
Defined gossip class.
Each nodes has a reference to a gossip object.
Gossip sends a message to a subset of random peers. When a gossip message is received, it checks the cache to see if it has been seen before. If it has, its ignored, otherwise the message is processed and forwarded.

Right now I'm only triggering gossip from main. We need to have nodes trigger gossip on alerts of bad/good data or connections.

This can also be extended for other use cases be on reputation, just need to set up a new message proto and handler callback. All the rest should stay consistent.